Testing

+++ src/main/java/nl/novi/eindopdrachtv3/services/ExamServiceImpl.java
@@ -33,7 +33,12 @@ public class ExamServiceImpl implements ExamService {
         List<ExamDto> edtoList = new ArrayList<>();
 
         for(Exam e : el) {
-            ExamDto edto = new ExamDto(e.getId(), e.getWrongEntries(), e.isPassed(), e.getTimestamp(), e.getWordList(), e.getUserProfile());
+            ExamDto edto = new ExamDto(
+                    e.getId(),
+                    e.getWrongEntries(),
+                    e.isPassed(), e.getTimestamp(),
+                    e.getWordList(),
+                    e.getUserProfile());
             edtoList.add(edto);
         }
         return edtoList;
@@ -66,7 +71,10 @@ public class ExamServiceImpl implements ExamService {
     }
 
     @Override
-    public void deleteExamById(Long id) {
+    public void deleteExamById(Long id) throws RecordNotFoundException {
+        if(!examRepository.findById(id).isEmpty()) {
+            throw new RecordNotFoundException("exam with id: " + id + " not present");
+        }
         examRepository.deleteById(id);
     }
